+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 - Added interactable volume slider
 - Added entity hints volume slider
-- Enemy hint sounds, jump hint sounds and hint sounds are now pitched down when they're behind the player
+- Enemy hint sounds, jump hint sounds, entity hint sounds and hint sounds are now pitched down when they're behind the player
 
 ### Fixed